api/requirements.txt
alembic==1.14.1
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cors==1.0.1
cryptography==44.0.0
dnspython==2.7.0
filelock==3.17.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
future==1.0.0
gevent==24.11.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jwcrypto==1.5.6
Mako==1.3.8
MarkupSafe==3.0.2
mongoose==0.0.1
mysql-connector==2.2.9
mysql-connector-python==9.2.0
pycparser==2.22
PyJWT==2.10.1
pymongo==4.11
PySocks==1.7.1
python-dotenv==1.0.1
python-jwt==4.1.0
python-wrap==0.1.0
requests==2.32.3
requests-file==2.1.0
setuptools==75.8.0
SQLAlchemy==2.0.37
tldextract==5.1.3
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
zope.event==5.0
zope.interface==7.2
